+++ oalg-base/CHANGELOG.md
@@ -69,3 +69,8 @@ has been defined by EmbeddingFrom.
 ## [2.0.0.2] - 2024-09-22
 ### Changed
 - data type Closer in module OAlg.Data.Ord changed to Closure.
+
+## [2.0.0.3] - 2024-10-01
+### Added
+- FSequence for total sequences with finite support.
+- PTree for efficient retrieving values from a partially defiend sequence.